+++ changelog.md
@@ -401,3 +401,11 @@
     * Added farm_add_rds_servers
     * Added farm_cancel_scheduled_maintenance
     * Added farm_remove_rds_servers
+
+06-01-2022
+* Inventory Class
+    * Added new_application_pool_v2
+    * Added update_application_pool_v2
+    * Added update_farm_v2
+    * Added add_global_desktop_entitlement_v2
+    * Added new_farm_v2